Implements changes proposed in feeedback. There might be something wrong as RSA plot is quite far from one in materials.
